feat: separated app file
