+++ src/index.ts
@@ -1,33 +1,13 @@
-import bodyParser from 'body-parser'
-import cors from 'cors'
 import dotenv from 'dotenv'
-import express from 'express'
-import helmet from 'helmet'
-import swaggerUi from 'swagger-ui-express'
-
-import swaggerOutput from './docs/api.json'
+import app from './app'
 import logger from './config/logger.config'
-import httpLogger from './config/httpLogger.config'
-import rateLimiterMiddleware from './middlewares/rateLimiter.middleware'
-import router from './routes'
-import { connectDatabase } from './database/dbconnect'
 
 dotenv.config()
 
-const app = express()
 const port = process.env.PORT ?? 3000
-
-app.use(httpLogger)
-app.use(bodyParser.json({ limit: '1mb' }))
-app.use(cors())
-app.use(helmet())
-app.use(rateLimiterMiddleware)
-
-app.use('/api', router)
-app.use('/docs', swaggerUi.serve, swaggerUi.setup(swaggerOutput))
-
-connectDatabase()
 
 app.listen(port, () => {
 	logger.debug(`running at http://localhost:${port}`, { port })
 })
+
+
